@@ -1,10 +1,10 @@
 <script>
-  let count = 0
+  let count = $state(0)
   const increment = () => {
     count += 1
   }
 </script>
 
-<button on:click={increment}>
+<button onclick={increment}>
   count is {count}
 </button>
