@@ -1,5 +1,5 @@
 import { render } from 'preact'
-import { App } from './app.tsx'
 import './index.css'
+import { App } from './app.tsx'
 
 render(<App />, document.getElementById('app')!)
