@@ -1,5 +1,5 @@
 import { render } from 'preact'
-import { App } from './app.jsx'
 import './index.css'
+import { App } from './app.jsx'
 
 render(<App />, document.getElementById('app'))
