@@ -1,7 +1,7 @@
 import '@builder.io/qwik/qwikloader.js'
 
 import { render } from '@builder.io/qwik'
-import { App } from './app.tsx'
 import './index.css'
+import { App } from './app.tsx'
 
 render(document.getElementById('app') as HTMLElement, <App />)
