@@ -1,7 +1,7 @@
 import '@builder.io/qwik/qwikloader.js'
 
 import { render } from '@builder.io/qwik'
-import { App } from './app.jsx'
 import './index.css'
+import { App } from './app.jsx'
 
 render(document.getElementById('app'), <App />)
