@@ -1,8 +1,7 @@
 /* @refresh reload */
 import { render } from 'solid-js/web'
-
 import './index.css'
-import App from './App'
+import App from './App.jsx'
 
 const root = document.getElementById('root')
 
